Lesta_task/requirements.txt
Django
python-dotenv
matplotlib
psycopg2-binary
djangorestframework==3.16.0
drf-yasg==1.21.10
